+++ src/ReactRendererFactory/index.ts
@@ -1,4 +1,4 @@
-import { Inject, Optional, Renderer2, RendererFactory2, RendererType2, Type } from '@angular/core';
+import { Inject, Optional, Renderer2, RendererFactory2, RendererType2, NgZone } from '@angular/core';
 import { CustomComponents } from '../types';
 import { ReactRenderer } from '../ReactRenderer';
 
@@ -7,12 +7,13 @@ export class ReactRendererFactory extends RendererFactory2 {
 
   constructor(
     @Inject('Context') ctx: any,
+    zone: NgZone,
     @Optional()
     @Inject('ElementSelector')
     selector: CustomComponents,
   ) {
     super();
-    this.renderer = new ReactRenderer(ctx, selector);
+    this.renderer = new ReactRenderer(ctx, zone, selector);
   }
 
   createRenderer(_hostElement: any, _type: RendererType2 | null): Renderer2 {
@@ -22,7 +23,7 @@ export class ReactRendererFactory extends RendererFactory2 {
   begin() {}
 
   end() {
-    this.renderer.render();
+    this.renderer.notify();
   }
 
   whenRenderingDone(): Promise<any> {