+++ src/index.ts
@@ -1,4 +1,4 @@
-import { ClassProvider, RendererFactory2 } from '@angular/core';
+import { NgModule, ClassProvider, RendererFactory2 } from '@angular/core';
 import { ReactRenderer } from './ReactRenderer';
 import { ReactRendererFactory } from './ReactRendererFactory';
 
@@ -11,3 +11,14 @@ export const Rectangly: ClassProvider = {
   provide: RendererFactory2,
   useClass: ReactRendererFactory,
 };
+
+export function createModuleBasedOn(baseModule: any) {
+  @NgModule({
+    imports: [baseModule],
+    exports: [baseModule],
+    providers: [Rectangly],
+  })
+  class RectanglyModule {}
+
+  return RectanglyModule;
+}