+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # rectangly
 
+## 0.2.0
+
+- Added a module creator
+- Run the callback listeners within NgZone
+- Rendering performed in dedicated React tree
+
 ## 0.1.1
 
-* Unit tests added
+- Unit tests added
 
 ## 0.1.0
 
-* Initial release
+- Initial release